feat(storybooks): build storybooks for CI and deploy to mozilla-fxa/storybooks

+++ _scripts/build-storybooks.sh
@@ -1,47 +1,57 @@
 #!/bin/bash -ex
 
-# TODO: only build if the current commit has changes to packages with storybooks - see modules-to-test.js for some hints?
+# Bail if we don't have a Github token - witheld for PRs from personal forks
+if [ -z $STORYBOOKS_GITHUB_TOKEN ]; then
+    echo "Github Personal Access Token required, skipping Storybook build"
+    exit;
+fi
 
 # Config from env vars with defaults
 SKIP_STORYBOOK_BUILD="${SKIP_STORYBOOK_BUILD:-false}"
 PROJECT_REPO="${PROJECT_REPO:-mozilla/fxa}"
-STORYBOOKS_REPO="${STORYBOOKS_REPO:-lmorchard/fxa-storybooks}"
-STORYBOOKS_URL="${STORYBOOKS_URL:-https://lmorchard.github.io/fxa-storybooks/}"
+STORYBOOKS_REPO="${STORYBOOKS_REPO:-mozilla-fxa/storybooks}"
+STORYBOOKS_URL="${STORYBOOKS_URL:-https://mozilla-fxa.github.io/storybooks}"
 STORYBOOKS_BRANCH="${STORYBOOKS_BRANCH:-gh-pages}"
 PUBLISH_ROOT="${PUBLISH_ROOT:-storybooks-publish}"
 MAX_STORYBOOK_AGE="${MAX_STORYBOOK_AGE:-90}"
 
+STORYBOOKS_REPO_URL="https://$STORYBOOKS_GITHUB_TOKEN@github.com/$STORYBOOKS_REPO.git"
+
 # Context for current build
 COMMIT_HASH=$(git rev-parse HEAD)
 
-# Build all the storybooks
+# Build all the available storybooks relevant to packages/test.list
 if [ "$SKIP_STORYBOOK_BUILD" == false ]; then
-    # fxa-payments-server relies on fxa-content-server .scss styles, which in turn
-    # rely on some modules in package.json
-    pushd ./packages/fxa-content-server
-    npm ci
-    popd
-
     STORYBOOKS=$(find packages -maxdepth 2 -type d -name '.storybook');
     for STORYBOOK_PATH in $STORYBOOKS; do
-        pushd $(dirname $STORYBOOK_PATH);
-        rm -rf storybook-static;
-        npm ci;
-        npm run build-storybook;
-        popd
+        PACKAGE_DIR=$(dirname $STORYBOOK_PATH)
+        PACKAGE_NAME=$(basename $PACKAGE_DIR)
+        if grep -e "$PACKAGE_NAME" -e 'all' "./packages/test.list" > /dev/null; then
+            pushd $PACKAGE_DIR;
+            rm -rf storybook-static;
+            yarn workspaces focus $PACKAGE_NAME
+            yarn run build-storybook;
+            popd
+        fi;
     done;
+fi
+
+# Check if any storybooks were built, bail if not
+if [[ -z $(find ./packages -type d -name 'storybook-static') ]]; then
+    echo "No Storybooks built, skipping the rest of deployment."
+    exit 0;
 fi
 
 # Get a fresh checkout of the repo to which we'll publish
 rm -rf $PUBLISH_ROOT
-git clone --branch $STORYBOOKS_BRANCH git@github.com:$STORYBOOKS_REPO.git $PUBLISH_ROOT
+git clone --branch $STORYBOOKS_BRANCH $STORYBOOKS_REPO_URL $PUBLISH_ROOT
 mkdir -p $PUBLISH_ROOT/{commits,pulls}
 
 # Copy all the storybooks into the publish directory for commit.
 COMMIT_PATH="$PUBLISH_ROOT/commits/$COMMIT_HASH"
 rm -rf $COMMIT_PATH
 mkdir -p $COMMIT_PATH
-for STORYBOOK_DIR in $(find . -type d -name 'storybook-static'); do
+for STORYBOOK_DIR in $(find ./packages -type d -name 'storybook-static'); do
     PACKAGE_NAME=$(basename $(dirname $STORYBOOK_DIR));
     BUILD_PATH="$COMMIT_PATH/$PACKAGE_NAME";
     cp -r $STORYBOOK_DIR $BUILD_PATH
@@ -71,7 +81,7 @@ fi;
 find $PUBLISH_ROOT/commits $PUBLISH_ROOT/pulls -mindepth 1 -maxdepth 1 \
     -type d -ctime +$MAX_STORYBOOK_AGE -exec rm -rf {} \;
 
-# Build index.html for root and commits
+# Build index.html for root, commits, and pulls
 ./_scripts/build-storybooks-indexes.js $PUBLISH_ROOT
 
 # Check for changes
@@ -89,6 +99,18 @@ git config user.email "fxa-core@mozilla.com"
 git add -A .
 git commit -qm "chore(docs): rebuild storybooks for ${COMMIT_HASH} [skip ci]"
 git push -q origin gh-pages
+
+# If this is a pull request, post a comment to the PR linking to the storybooks
+if [[ ! -z $CI_PULL_REQUEST ]]; then
+    PR_NUMBER=$(echo $CI_PULL_REQUEST | cut -d/ -f7);
+    PR_URL="$STORYBOOKS_URL/pulls/$PR_NUMBER/";
+    COMMIT_URL="$STORYBOOKS_URL/commits/$COMMIT_HASH/";
+    curl \
+        -H'Content-Type: application/json' \
+        -H"Authorization: token $STORYBOOKS_GITHUB_TOKEN" \
+        --data "{\"body\":\"# Storybook Deployment\\nPull Request: $PR_URL\\nCommit: $COMMIT_URL\"}" \
+        https://api.github.com/repos/${PROJECT_REPO}/issues/${PR_NUMBER}/comments;
+fi
 
 echo "Cleaning up $PUBLISH_ROOT"
 cd ..
